Avoid sending event code when enabling soft events.
The flag that indicates whether software events are enabled and the
actual event code to be sent for a software event share the same
register.

This meant that enabling software events would have the effect of
sending the event code that was sent last, because the device sends out
the event code when writing to the register and the flag is enabled.

This commit fixes this issue by ensuring that the event code part of the
register is set to zero when enabling software events.