Add info about delays and pulse widths in microseconds.
In addition to showing the settings in event clock cycles (how they are
configured), we also show them in microseconds. This calculation is only
correct if the PV containing the event clock frequency (EventClock:Freq)
has been configured correctly and the caculated values are not precise
because the event clock frequency is only known in full units of MHz.